Completed plumbing code for new SV fields.

+++ hmf-common/src/main/java/com/hartwig/hmftools/common/variant/structural/StructuralVariantData.java
@@ -35,12 +35,12 @@ public abstract class StructuralVariantData {
     public abstract boolean imprecise();
     public abstract Integer somaticScore();
     public abstract Double qualityScore();
-    public abstract Long startIntervalOffsetStart();
-    public abstract Long startIntervalOffsetEnd();
-    public abstract Long endIntervalOffsetStart();
-    public abstract Long endIntervalOffsetEnd();
-    public abstract Long inexactHomologyOffsetStart();
-    public abstract Long inexactHomologyOffsetEnd();
+    public abstract Integer startIntervalOffsetStart();
+    public abstract Integer startIntervalOffsetEnd();
+    public abstract Integer endIntervalOffsetStart();
+    public abstract Integer endIntervalOffsetEnd();
+    public abstract Integer inexactHomologyOffsetStart();
+    public abstract Integer inexactHomologyOffsetEnd();
     public abstract Integer startTumourVariantFragmentCount();
     public abstract Integer startTumourReferenceFragmentCount();
     public abstract Integer startNormalVariantFragmentCount();
